requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
openai==0.27.0
pytest==7.3.1
gunicorn==20.1.0
